restructuring of how we think of navigation and page wrappers; global error simplification
