+++ src/Components/Presentational/ErrorSplashes/BadRequest.js
@@ -1,0 +1,14 @@
+// TODO: Once Scott's error splashes land, we can implement them each as their
+// own simple components in this folder
+import React from 'react';
+
+const BadRequest = () => {
+  return (
+    <>
+      it appears your request was bad! if you're seeing this, it could be due to
+      a failed auth thing. ope.
+    </>
+  );
+};
+
+export default BadRequest;
