+++ src/Components/Scenes/Chrome/PageWrapper/PageWrapper-Container.js
@@ -1,20 +1,11 @@
 import { connect } from 'react-redux';
 import PageWrapper from './PageWrapper';
-import { authenticateFetch } from '../../../../State/AuthFlow/actions';
 
-const mapStateToProps = ({ searchResult, currentUser }) => {
+const mapStateToProps = ({ searchResult }) => {
   return {
     isSearchClicked: searchResult.isSearchClicked,
-    isSearchClosed: searchResult.isSearchClosed,
-    sessionToken: currentUser.SessionToken,
-    error: currentUser.error
+    isSearchClosed: searchResult.isSearchClosed
   };
 };
 
-const mapDispatchToProps = dispatch => {
-  return {
-    authenticateFetch: token => dispatch(authenticateFetch(token))
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(PageWrapper);
+export default connect(mapStateToProps)(PageWrapper);
