+++ src/Components/Scenes/Chrome/PageWrapper/PageWrapper.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Route, useLocation } from 'react-router-dom';
 import { AnimatedSwitch } from 'react-router-transition';
 import { makeStyles } from '@material-ui/core/styles';
@@ -11,8 +11,6 @@ import SearchResultsContainer from '../../SearchResults/SearchResults/SearchResu
 import RequestAssetContainer from '../../RequestAssets/RequestAsset/RequestAsset-Container';
 import SentRequestsContainer from '../../ViewRequests/SentRequests/SentRequests-Container';
 import NewJobContainer from '../../Hydration/NewJob/NewJob/NewJob-Container';
-import { useQuery } from '../../../../Hooks/customHooks';
-import { isEmptyObject } from '../../../../Utilities/utils';
 
 const pageContainerStyle = makeStyles(theme => ({
   pageContainer: {
@@ -28,37 +26,12 @@ const pageContainerStyle = makeStyles(theme => ({
 }));
 
 const PageWrapper = props => {
-  const { isSearchClicked, sessionToken, error, authenticateFetch } = props;
+  const { isSearchClicked } = props;
   const classes = pageContainerStyle();
-  const samlResponse = useQuery('SAMLResponse');
   const curPath = useLocation().pathname;
 
-  const [isRedirecting, setIsRedirecting] = useState(true);
-  // This functionality could be useful if wrapped in an error view button or link,
-  // so in anticipation, it's set up as a function for now
-  const redirect = () => {
-    setIsRedirecting(true);
-    window.location.replace('https://pennymac.onelogin.com/portal/');
-  };
-
-  // Case 1: there is a SAML response but no session token, so authenticate real quick
-  // Case 2: there is neither a SAML response nor a session token, so redirect to OneLogin
-  // Case 3: there may or may not be a SAML response,
-  //         but there is a session token, so relax until an hour later
-  //         when we get a 4xx code from some request, then redirect
-  useEffect(() => {
-    if (samlResponse && !sessionToken) {
-      setIsRedirecting(false);
-      authenticateFetch(samlResponse);
-    } else if (!sessionToken) {
-      redirect();
-    } else {
-      setIsRedirecting(false);
-    }
-  }, [samlResponse, authenticateFetch, sessionToken]);
-
-  const authenticatedContent = (
-    <>
+  return (
+    <div className={classes.pageContainer}>
       {curPath === '/hydration/new-job' ? null : <AppBarContainer />}
 
       <AnimatedSwitch
@@ -82,28 +55,8 @@ const PageWrapper = props => {
       </AnimatedSwitch>
 
       {isSearchClicked ? <SearchContainer /> : null}
-    </>
-  );
-
-  // TODO: replace <span> with proper error splash once it's landed
-  return (
-    <div className={classes.pageContainer}>
-      {!isRedirecting ? (
-        isEmptyObject(error) ? (
-          authenticatedContent
-        ) : (
-          <span>auth error placeholder</span>
-        )
-      ) : null}
     </div>
   );
-
-  // Uncomment to ignore auth
-  // return (
-  //   <div className={classes.pageContainer}>
-  //     authenticatedContent
-  //   </div>
-  // );
 };
 
 export default PageWrapper;
