+++ src/Reducers/initialState.js
@@ -3,7 +3,6 @@ import { initialState as viewRequests } from '../State/ViewRequests/reducers';
 const initialState = {
   // TODO: Replace Stub Data (role and EmpEmail) with Real Data
   currentUser: {
-    error: {},
     role: {
       governance: true,
       approver: true
